[EDIFIKANA] #388 updating tests, changing Long to Double for currency values,

+++ edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/datastore/RentConfigDatastore.kt
@@ -26,7 +26,7 @@ interface RentConfigDatastore {
      */
     suspend fun setRentConfig(
         unitId: UnitId,
-        monthlyAmount: Long,
+        monthlyAmount: Double,
         dueDay: Int,
         currency: String,
         updatedBy: UserId?,
